sched: switch CR3 after ESP in the tick stub, not inside schedule()
schedule() ran change_page_directory() while still executing on the outgoing
thread's kernel stack. For kernel threads that is fine (their stacks live at
0x401000+, PDE 1, cloned into every address space), but a ring-3 process's
kernel stack sits at ~0x800000+ (PDE 2), which is not shared -- so scheduling
such a process away (typically on exit, in the remove_proc path) pulled the
stack out from under the rest of schedule() and the asm epilogue: #PF in
load_pdbr's `pop %ebp` right after `mov %cr3`, CR2 == ESP, -> #DF -> triple
fault. A race: it passed a 25/25 soak but hit ~50% once teardown timing shifted.

schedule() now returns a packed 64-bit value -- resume ESP in the low half,
the CR3 to load (or 0) in the high half -- and does not touch CR3 itself. The
LAPIC timer / resched stubs switch ESP first (a register move, no memory
access), then load the new CR3, so the new stack is valid before it is read.

Soak: 12x hda/mem + 12x hda/hello interleaved, hda/{example,lua} -- no faults.

+++ sched.h
@@ -45,12 +45,17 @@ process_t *get_cur_proc();
 /** @return The process owning thread id @p id, or NULL. */
 process_t *get_proc_by_id(int id);
 /**
- * @brief Timer-IRQ scheduler tick: save @p esp, pick the next runnable
- *        process/thread and return its kernel stack pointer.
+ * @brief Timer/IPI scheduler tick: save @p esp, pick the next runnable
+ *        process/thread and hand the asm stub what it needs to resume it.
+ *
  * @param esp Kernel ESP of the interrupted thread.
- * @return Kernel ESP to resume on.
+ * @return A packed value: the low 32 bits are the kernel ESP to resume on; the
+ *         high 32 bits are the page-directory (CR3) to load, or 0 to keep the
+ *         current one. The caller must load CR3 *after* switching ESP — the
+ *         outgoing thread's kernel stack may not be mapped in the new address
+ *         space, and this one's may not be mapped in the old one.
  */
-uint32_t schedule(uint32_t esp);
+uint64_t schedule(uint32_t esp);
 /**
  * @brief Voluntarily give up the CPU.
  *
